Adds some options to meta Renamed api->root where all api paths join Upgraded openapi overrides since last fastapi interface changed
