+++ services/api-server/requirements/_base.in
@@ -9,11 +9,13 @@ fastapi[all]
 aiopg[sa]
 tenacity
 passlib[bcrypt]
-pyjwt
 loguru
 pydantic[dotenv]
 cryptography
 httpx
+
+# TODO: check alternative https://github.com/latchset/jwcrypto/
+pyjwt>=1.7.1  # Vulnerable SEE https://auth0.com/blog/critical-vulnerabilities-in-json-web-token-libraries/?_ga=2.21160507.1609921856.1592236287-1918774871.1591379535
 
 
 async-exit-stack  # not needed when python>=3.7
